Feat : modal,alert 폴더명 및 파일명 변경점 모든 페이지 적용
