+++ src/components/molecules/Alert/Alert.jsx
@@ -1,74 +1,32 @@
-import { useNavigate } from 'react-router-dom';
-import { OpacityBg } from '../Modal/styled';
-import AlertWrapper from './styled';
+import { useRecoilState } from 'recoil';
+import { alertAtom, modalAtom } from '../../../recoil/atom';
+import NewAlertModalWrapper from './styled';
 
-const Alert = ({
-  onClose,
-  type,
-  questionText,
-  rightBtnText,
-  onClickDeletePost,
-  onClickReportPost,
-  onClickDeleteComment,
-  onClickReportComment,
-  postId,
-  commentId,
-}) => {
-  const navigate = useNavigate();
+const Alert = ({ onClickAlertEventHandler }) => {
+  const [modal, setModal] = useRecoilState(modalAtom);
+  const [alert, setAlert] = useRecoilState(alertAtom);
 
-  // 로그아웃
-  const removeInfo = () => {
-    window.localStorage.removeItem('token');
-    window.localStorage.removeItem('accountname');
-    navigate('/');
-  };
-
-  const onClickDeleteBtn = () => {
-    // 내 댓글 삭제
-    if (type === 'comment' && rightBtnText === '삭제') {
-      onClickDeleteComment(postId, commentId);
-    }
-
-    // 댓글 신고
-    if (type === 'comment' && rightBtnText === '신고') {
-      onClickReportComment(postId, commentId);
-      onClose();
-    }
-
-    // 피드 게시글 삭제
-    if (type === 'feedpost' && rightBtnText === '삭제') {
-      onClickDeletePost(postId);
-      onClose();
-      navigate(`/myprofile`);
-    }
-
-    // 피드 게시글 신고
-    if (type === 'feedpost' && rightBtnText === '신고') {
-      onClickReportPost(postId);
-      onClose();
-    }
+  const onClickCancelModal = () => {
+    setModal({
+      ...modal,
+      isActive: { header: false, post: false, comment: false },
+    });
+    setAlert({
+      ...alert,
+      isActive: { header: false, post: false, comment: false },
+    });
   };
 
   return (
-    <OpacityBg>
-      <AlertWrapper>
-        <div>
-          <strong>{questionText}</strong>
-          <button type='button' onClick={onClose}>
-            취소
-          </button>
-          {rightBtnText === '로그아웃' ? (
-            <button type='button' onClick={removeInfo} className='right'>
-              {rightBtnText}
-            </button>
-          ) : (
-            <button type='button' onClick={onClickDeleteBtn} className='right'>
-              {rightBtnText}
-            </button>
-          )}
-        </div>
-      </AlertWrapper>
-    </OpacityBg>
+    <NewAlertModalWrapper>
+      <p>{alert.text.alertText}</p>
+      <button type='button' onClick={onClickCancelModal}>
+        취소
+      </button>
+      <button type='button' onClick={onClickAlertEventHandler}>
+        {alert.text.text}
+      </button>
+    </NewAlertModalWrapper>
   );
 };
