+++ src/components/molecules/Alert/styled.jsx
@@ -1,50 +1,43 @@
 import styled, { css } from 'styled-components';
 
-const AlertWrapper = styled.article`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  border-radius: 10px;
-  width: 260px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
+const NewAlertModalWrapper = styled.article`
   ${({ theme }) => {
     return css`
+      width: 252px;
+      height: 120px;
       background-color: ${theme.colors.colorBg};
       color: ${theme.colors.colorBorder};
+      position: absolute;
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%);
+      border-radius: 10px;
+      z-index: 9999;
+      display: flex;
+      flex-wrap: wrap;
+
+      p {
+        width: 100%;
+        border-bottom: 1px solid ${theme.colors.colorBorder};
+        font-size: ${theme.fontSizes.xl};
+        font-weight: 500;
+        line-height: 73px;
+        text-align: center;
+      }
+      button {
+        width: 126px;
+        height: 46px;
+        display: block;
+        border-radius: 0 0 0 10px;
+        color: ${theme.colors.colorBorder};
+
+        &:last-child {
+          border-radius: 0 0 10px 0;
+          border-left: 1px solid ${theme.colors.colorBorder};
+          color: ${theme.colors.colorWarning};
+        }
+      }
     `;
   }}
-
-  div {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    justify-items: center;
-    align-items: center;
-  }
-
-  strong {
-    margin: 22px 0;
-    grid-column: 1/3;
-  }
-
-  button {
-    width: 126px;
-    height: 46px;
-    ${({ theme }) => {
-      return css`
-        color: ${theme.colors.colorBorder};
-        border-top: 1px solid ${theme.colors.colorDB};
-        font-size: ${theme.fontSizes.base};
-        &.right {
-          color: ${theme.colors.colorWarning};
-          border-left: 1px solid ${theme.colors.colorDB};
-        }
-      `;
-    }}
-  }
 `;
-
-export default AlertWrapper;
+export default NewAlertModalWrapper;
